+++ HISTORY.md
@@ -1,3 +1,13 @@
+## 7.0.0
+
+* `CHANGED` update `uxcore-date-form-field` to `^0.9.0`
+* `CHANGED` update `uxcore-form-field` to `^0.3.0`
+* `CHANGED` update `uxcore-input-form-field` to `^0.3.0`
+* `CHANGED` update `uxcore-pickable-form-field` to `^0.5.0`
+* `CHANGED` update `uxcore-select2` to `^0.5.0`
+* `CHANGED` update `uxcore-textarea-form-field` to `^0.2.0`
+* `CHANGED` adapt React@16
+
 ## 6.3.4
 
 * `FIXED` SearchField handleDataChange